Multilayered comms between client & server w/ projects.
